common/refactor
