+++ cmd/build.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"path/filepath"
 	"plenti/cmd/build"
+	"plenti/common"
 	"plenti/readers"
 	"time"
 
@@ -91,12 +92,8 @@ func Build() {
 	// Clear out any previous build dir of the same name.
 	if _, buildPathExistsErr := os.Stat(buildPath); buildPathExistsErr == nil {
 		build.Log("Removing old '" + buildPath + "' build directory")
-		err := os.RemoveAll(buildPath)
 
-		if err != nil {
-			log.Fatal(err)
-
-		}
+		common.CheckErr(os.RemoveAll(buildPath))
 	}
 
 	// Create the buildPath directory.
@@ -108,12 +105,12 @@ func Build() {
 	build.Log("Creating '" + buildDir + "' build directory")
 
 	// Add core NPM dependencies if node_module folder doesn't already exist.
-	if err = build.NpmDefaults(tempBuildDir); err != nil {
-		log.Fatal(err)
-	}
+	err = build.NpmDefaults(tempBuildDir)
+	common.CheckErr(err)
 
 	// Write ejectable core files to filesystem before building.
-	tempFiles, ejectedPath := build.EjectTemp(tempBuildDir)
+	tempFiles, ejectedPath, err := build.EjectTemp(tempBuildDir)
+	common.CheckErr(err)
 
 	// Directly copy .js that don't need compiling to the build dir.
 	if err = build.EjectCopy(buildPath, tempBuildDir, ejectedPath); err != nil {
@@ -124,44 +121,36 @@ func Build() {
 	//bundledContent := build.Bundle()
 
 	// Directly copy static assets to the build dir.
-	if err := build.AssetsCopy(buildPath, tempBuildDir); err != nil {
-		log.Fatal(err)
-	}
+	common.CheckErr(build.AssetsCopy(buildPath, tempBuildDir))
 
 	// Run the build.js script using user local NodeJS.
 	if NodeJSFlag {
-		clientBuildStr := build.NodeClient(buildPath)
+		clientBuildStr, err := build.NodeClient(buildPath)
+		common.CheckErr(err)
 		staticBuildStr, allNodesStr, err := build.NodeDataSource(buildPath, siteConfig)
-		if err != nil {
-			log.Fatal(err)
-		}
+		common.CheckErr(err)
 
-		if err := build.NodeExec(clientBuildStr, staticBuildStr, allNodesStr); err != nil {
-			log.Fatal(err)
-		}
+		common.CheckErr(build.NodeExec(clientBuildStr, staticBuildStr, allNodesStr))
 	} else {
 
 		// Prep the client SPA.
-		if err := build.Client(buildPath, tempBuildDir, ejectedPath); err != nil {
-			log.Fatal(err)
-		}
+
+		common.CheckErr(build.Client(buildPath, tempBuildDir, ejectedPath))
 
 		// Build JSON from "content/" directory.
-		if err := build.DataSource(buildPath, siteConfig, tempBuildDir); err != nil {
-			log.Fatal(err)
-		}
+		common.CheckErr(build.DataSource(buildPath, siteConfig, tempBuildDir))
 
 	}
 
 	// Run Gopack (custom Snowpack alternative) for ESM support.
-	build.Gopack(buildPath)
+	common.CheckErr(build.Gopack(buildPath))
 
 	if tempBuildDir != "" {
 		// If using themes, just delete the whole build folder.
-		CheckErr(build.ThemesClean(tempBuildDir))
+		common.CheckErr(build.ThemesClean(tempBuildDir))
 	} else {
 		// If no theme, just delete any ejectable files that the user didn't manually eject.
-		CheckErr(build.EjectClean(tempFiles, ejectedPath))
+		common.CheckErr(build.EjectClean(tempFiles, ejectedPath))
 	}
 
 }
