+++ common/errs.go
@@ -1,4 +1,4 @@
-package cmd
+package common
 
 import "log"
